Reduce duplicate code
