+++ test/Altinn.Notifications.IntegrationTests/Notifications/InstantOrdersController/InstantOrdersControllerTests.cs
@@ -8,13 +8,8 @@ using Altinn.Common.AccessToken.Services;
 using Altinn.Notifications.Controllers;
 using Altinn.Notifications.Core.Configuration;
 using Altinn.Notifications.Core.Integrations;
-using Altinn.Notifications.Core.Models;
-using Altinn.Notifications.Core.Models.Address;
-using Altinn.Notifications.Core.Models.NotificationTemplate;
 using Altinn.Notifications.Core.Models.Orders;
-using Altinn.Notifications.Core.Models.ShortMessageService;
 using Altinn.Notifications.Core.Services.Interfaces;
-using Altinn.Notifications.Core.Shared;
 using Altinn.Notifications.Models.Orders;
 using Altinn.Notifications.Models.Recipient;
 using Altinn.Notifications.Tests.Notifications.Mocks.Authentication;
